a series of updates

Former-commit-id: d6b99f2f050af77c7d3974fa737c221a88c36e4e

+++ src/metadata.hh
@@ -33,25 +33,25 @@ struct Metadata {
 };
 
 typedef struct LorenzoZipContext {
-    int    d0, d1, d2;
-    int    stride1, stride2;
-    int    radius;
-    double ebx2_r;
+    int   d0, d1, d2;
+    int   stride1, stride2;
+    int   radius;
+    float ebx2_r;
 } lorenzo_zip;
 
 typedef struct LorenzoDryrunContext {
-    int    d0, d1, d2;
-    int    stride1, stride2;
-    int    radius;
-    double ebx2_r, ebx2;
+    int   d0, d1, d2;
+    int   stride1, stride2;
+    int   radius;
+    float ebx2_r, ebx2;
 } lorenzo_dryrun;
 
 typedef struct LorenzoUnzipContext {
-    int    d0, d1, d2;
-    int    nblk0, nblk1, nblk2;
-    int    stride1, stride2;
-    int    radius;
-    double ebx2;
+    int   d0, d1, d2;
+    int   nblk0, nblk1, nblk2;
+    int   stride1, stride2;
+    int   radius;
+    float ebx2;
 } lorenzo_unzip;
 
 template <int ndim>
